Update navigation header to avoid showing back arrow on homepage
Found a bug when playing around with this today: On mobile, navigate to page 2 by pressing "Next". Now navigate back to page 1 by pressing "Prev". The navigating header will be the back arrow. Clicking it will take you back to page 2, which makes no sense.

This should fix it, but for full transparency I did not test this and probably didn't format the code right, I edited it within github.com. Just wanted to at least draw attention to the issue.